model-training/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
pillow>=10.0.0
pandas>=2.0.0
requests>=2.31.0
tqdm>=4.66.0
scikit-learn>=1.3.0
huggingface_hub>=0.20.0
